requirements.txt
alembic==1.7.7
aniso8601==9.0.1
appdirs==1.4.4
audioread==2.1.9
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cycler==0.11.0
decorator==5.1.1
Flask==2.1.2
Flask-API==3.0.post1
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-Script==2.0.5
Flask-SQLAlchemy==2.5.1
fonttools==4.33.3
gcloud==0.18.3
google-api-core==2.7.3
google-auth==2.6.6
google-cloud-texttospeech==2.11.0
googleapis-common-protos==1.56.0
greenlet==1.1.2
grpcio==1.46.1
grpcio-status==1.46.1
httplib2==0.20.4
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
kiwisolver==1.4.2
librosa==0.9.1
llvmlite==0.38.0
Mako==1.2.0
MarkupSafe==2.1.1
matplotlib==3.5.2
noisereduce==2.0.1
numba==0.55.1
numpy==1.21.6
oauth2client==4.1.3
packaging==21.3
Pillow==9.1.0
pooch==1.6.0
proto-plus==1.20.3
protobuf==3.20.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
resampy==0.2.2
rsa==4.8
scikit-learn==1.1.0
scipy==1.8.0
six==1.16.0
SoundFile==0.10.3.post1
SQLAlchemy==1.4.36
threadpoolctl==3.1.0
tqdm==4.64.0
unicodedata2==14.0.0
urllib3==1.26.9
Werkzeug==2.1.2
